+++ init.js
@@ -14,8 +14,26 @@ function browserSupport() {
 function addImportMap() {
   const base = document.currentScript.src.replace(/\/init.js$/, `/`)
 
-  window.importmap = {
-    ...(window.importmap || {}),
+  let importmap = document.querySelector(`[type=importmap]`)
+
+  if (!importmap) {
+    importmap = Object.assign(document.createElement(`script`), {
+      type: `importmap`,
+      innerHTML: `{}`,
+    })
+    document.head.appendChild(importmap)
+  }
+
+  let json
+  try {
+    json = JSON.parse(importmap.innerHTML)
+  } catch (e) {
+    console.error(`Importmap does not contain valid JSON, overwriting it.`)
+    json = {}
+  }
+
+  json.imports = {
+    ...(json.imports || {}),
     tiden: base + "tiden.js",
     "tiden/": base,
     "lit-html": "https://cdn.jsdelivr.net/npm/lit-html@1.4.1/lit-html.js",
@@ -28,34 +46,21 @@ function addImportMap() {
     reselect: "https://cdn.jsdelivr.net/npm/reselect@^4.0.0/es/index.js",
   }
 
-  document.head.appendChild(
-    Object.assign(document.createElement("script"), {
-      type: "importmap",
-      innerHTML: JSON.stringify({
-        imports: window.importmap,
-      }),
-    })
-  )
+  importmap.innerHTML = JSON.stringify(json, null, 2)
 }
 
 function addImportMapShim() {
-  const shim = Object.assign(document.createElement("script"), {
-    type: "script",
-    src: "https://unpkg.com/es-module-shims@0.12.1/dist/es-module-shims.js",
-  })
-  shim.setAttribute(`async`, ``)
-  shim.dataset.initial = true
-  document.head.appendChild(shim)
+  document.write(
+    `<script data-initial async src="https://unpkg.com/es-module-shims@0.12.1/dist/es-module-shims.js"></script>`
+  )
 }
 
 function startApp() {
-  const init = Object.assign(document.createElement("script"), {
-    type: "module",
-    src: `${window.root || `./`}index.js`,
-  })
-  init.setAttribute("async", "")
-  init.dataset.initial = ""
-  document.head.appendChild(init)
+  document.write(
+    `<script async data-initial type="module" src="${
+      window.root || `./`
+    }index.js"></script>`
+  )
 }
 
 if (!browserSupport()) {