+++ ops/init.js
@@ -14,7 +14,6 @@ import manifestJson from "./init/manifest.json.template.js"
 import * as touchable from "./init/touchable.js"
 
 export default async function init({ name, description, isTest }) {
-
   const files = await fs.readdir(process.cwd())
   if (files.length > 0) {
     console.log(`Cannot init project in non-empty folder`)
@@ -54,6 +53,16 @@ export default async function init({ name, description, isTest }) {
   })
   await createNano({
     name: `home`,
+    body: o`
+      yield template(root, function*(root) {
+        
+      })
+    `,
+    imports: {
+      "./template.js": {
+        default: [`template`],
+      },
+    },
   })
   await createComponent({
     name: `touchable`,